Repository Restructure
Trying to make the repository easier to navigate.  Folders breaking up the libraries into sections with names to describe the contents.
